chore: rename directory to match with primary commponent
