+++ src/pages/UserSettings/content/formSections/WorkspaceMapPin.section.tsx
@@ -8,8 +8,8 @@ import { FlexSectionContainer } from './elements'
 import { MAP_GROUPINGS } from 'src/stores/Maps/maps.groupings'
 import { required } from 'src/utils/validators'
 import { randomIntFromInterval } from 'src/utils/helpers'
-import { MAX_PIN_LENGTH } from 'src/pages/Settings/constants'
-import { fields, headings } from 'src/pages/Settings/labels'
+import { MAX_PIN_LENGTH } from 'src/pages/UserSettings/constants'
+import { fields, headings } from 'src/pages/UserSettings/labels'
 
 import type { ThemeStore } from 'src/stores/Theme/theme.store'
 import type { ILocation } from 'src/models/common.models'
